% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contingency_dummies}
\alias{contingency_dummies}
\title{Dummys from contingency}
\usage{
contingency_dummies(tbl)
}
\arguments{
\item{tbl}{a contingency table}
}
\value{
/code{X} the matrix of dummy row categories observations and /code{Y}
the dummy of dummy column categories observations
}
\description{
Returns two matrices, one representing dummy observations
of rows and one representing dummy observations of colummns
}
