% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decay.R
\name{get_decay}
\alias{get_decay}
\title{get_decay}
\usage{
get_decay(n, halflife = 1, method = "semi-geometric")
}
\arguments{
\item{n}{size.}

\item{halflife}{halflife as a percentage of n.}

\item{method}{it can be geometric or 
semi-gemetric.}
}
\description{
Creates vector with decay factor.
}
\details{
See reference at 
http://stats.stackexchange.com/questions/20542/generating-a-3-month-half-life-weighting-series-in-r
it is desirable to find how both methods change with halflife of 0 and 1
}

