% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices.R
\name{get_prices_yahoo}
\alias{get_prices_yahoo}
\title{get_prices_yahoo}
\usage{
get_prices_yahoo(yahoo_id, from, to = Sys.Date(), adjust = TRUE)
}
\arguments{
\item{yahoo_id}{character array containining assets'
yahoo ids.}

\item{from}{start date.}

\item{to}{end date. Sys.Date() is the default value}

\item{boolean}{indicating whether adjusted prices
should be returned.}
}
\description{
Builds an xts containing asset prices
downloaded from http://ichart.finance.yahoo.com/.
}
\details{
xx
}

