% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_returns}
\alias{get_returns}
\title{get_returns}
\usage{
get_returns(x, is_prices = TRUE, roll = 1, halflife = 1,
  is_cumulative = FALSE)
}
\arguments{
\item{x}{xts containing prices or returns}

\item{is_prices}{boolean indicating whether x 
contains prices or returns}

\item{roll}{rolling window}

\item{halflife}{half life as percentage of roll}

\item{is_cumulative}{boolean indicating whether 
cumulative return should be computed}
}
\description{
Creates xts of returns
}
\details{
xx
}

