% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{plot_returns}
\alias{plot_returns}
\title{plot_returns}
\usage{
plot_returns(returns)
}
\arguments{
\item{returns}{xts containg daily or cumulative 
returns}
}
\description{
Plots returns
}
\details{
xx
}

