% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{senado_partidos}
\alias{senado_partidos}
\title{Partidos de los Senadores Mexicanos}
\format{\itemize{
    \item \code{SENADOR} vector de clase \code{character} con el nombre de cada senador
    \item \code{PARTIDO} vector de clase \code{character} con el partido durante el periodo
}}
\source{
Los datos se obtuvieron de \link{www.senado.gob.mx} usando
\link{github.com/mauriciogtec/votacionEspacialMexico}
}
\usage{
data(senado_partidos)
}
\description{
Datos de votaciones de la 63 legislatura de Mexico, primer agno, primer periodo ordinario.
}
\keyword{datasets}

