% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{senado_votaciones}
\alias{senado_votaciones}
\title{datos de votaciones del senado}
\format{primeras tres columnas informacion de votaciones
las siguientes votaciones de cada senador
1 es que voto que si}
\usage{
senado_votaciones
}
\description{
datos de votaciones del senado
}
\keyword{datasets}

