% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesFour.R
\name{timesFour}
\alias{timesFour}
\title{Multiplicar por cuatro}
\usage{
timesFour(x)
}
\arguments{
\item{x}{un elemento o vector numerico}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector \code{x} por cuatro.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento \code{x} que puede ser un numero o un vector numerico y lo devuelve
multiplicado por cuatro. La documentacion de ayuda es generada usando roxygen2. La implementacion utiliza \code{\link{timesTwo}}
}
\examples{
timesFour(3)
timesFour(0:10)
}
