% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesTwo.R
\name{timesTwo}
\alias{timesTwo}
\title{Multiplicar por dos}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{un elemento o vector numerico}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector \code{x} por dos.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento \code{x} que puede ser un numero o un vector numerico y lo devuelve
multiplicado por dos. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
timesTwo(3)
timesTwo(0:10)
}

