% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{weighted.geometric.mean}
\alias{weighted.geometric.mean}
\title{weighted.geometric.mean}
\usage{
weighted.geometric.mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a vector or dataframe containing the values
whose weighted geometric mean is to be computed.}

\item{w}{a vector of weights.}

\item{na.rm}{remove NA for values before processing,
the default value is FALSE.}
}
\description{
Compute a weighted geometric mean.
}
\details{
See weighted.mean \{stats\} or geometric.mean 
\{psych\}.
}

